+++ pubsub/subscriptions/async_pull.go
@@ -1,0 +1,70 @@
+// Copyright 2019 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// Package subscription is a tool to manage Google Cloud Pub/Sub subscriptions by using the Pub/Sub API.
+// See more about Google Cloud Pub/Sub at https://cloud.google.com/pubsub/docs/overview.
+package subscription
+
+import (
+	"context"
+	"fmt"
+	"sync"
+
+	"cloud.google.com/go/pubsub"
+)
+
+func pullMsgs(client *pubsub.Client, subName string, topic *pubsub.Topic) error {
+	ctx := context.Background()
+
+	// Publish 10 messages on the topic.
+	var results []*pubsub.PublishResult
+	for i := 0; i < 10; i++ {
+		res := topic.Publish(ctx, &pubsub.Message{
+			Data: []byte(fmt.Sprintf("hello world #%d", i)),
+		})
+		results = append(results, res)
+	}
+
+	// Check that all messages were published.
+	for _, r := range results {
+		_, err := r.Get(ctx)
+		if err != nil {
+			return err
+		}
+	}
+
+	// [START pubsub_subscriber_async_pull]
+	// [START pubsub_quickstart_subscriber]
+	// Consume 10 messages.
+	var mu sync.Mutex
+	received := 0
+	sub := client.Subscription(subName)
+	cctx, cancel := context.WithCancel(ctx)
+	err := sub.Receive(cctx, func(ctx context.Context, msg *pubsub.Message) {
+		msg.Ack()
+		fmt.Printf("Got message: %q\n", string(msg.Data))
+		mu.Lock()
+		defer mu.Unlock()
+		received++
+		if received == 10 {
+			cancel()
+		}
+	})
+	if err != nil {
+		return err
+	}
+	// [END pubsub_subscriber_async_pull]
+	// [END pubsub_quickstart_subscriber]
+	return nil
+}
