+++ pubsub/subscriptions/create.go
@@ -12,40 +12,30 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Sample pubsub-quickstart creates a Google Cloud Pub/Sub topic.
-package main
+// Package subscription is a tool to manage Google Cloud Pub/Sub subscriptions by using the Pub/Sub API.
+// See more about Google Cloud Pub/Sub at https://cloud.google.com/pubsub/docs/overview.
+package subscription
 
-// [START pubsub_quickstart_create_topic]
+// [START pubsub_create_pull_subscription]
 import (
 	"context"
 	"fmt"
-	"log"
+	"time"
 
 	"cloud.google.com/go/pubsub"
 )
 
-func main() {
+func create(client *pubsub.Client, subName string, topic *pubsub.Topic) error {
 	ctx := context.Background()
-
-	// Sets your Google Cloud Platform project ID.
-	projectID := "YOUR_PROJECT_ID"
-
-	// Creates a client.
-	client, err := pubsub.NewClient(ctx, projectID)
+	sub, err := client.CreateSubscription(ctx, subName, pubsub.SubscriptionConfig{
+		Topic:       topic,
+		AckDeadline: 20 * time.Second,
+	})
 	if err != nil {
-		log.Fatalf("Failed to create client: %v", err)
+		return err
 	}
-
-	// Sets the name for the new topic.
-	topicName := "my-topic"
-
-	// Creates the new topic.
-	topic, err := client.CreateTopic(ctx, topicName)
-	if err != nil {
-		log.Fatalf("Failed to create topic: %v", err)
-	}
-
-	fmt.Printf("Topic %v created.\n", topic)
+	fmt.Printf("Created subscription: %v\n", sub)
+	return nil
 }
 
-// [END pubsub_quickstart_create_topic]
+// [END pubsub_create_pull_subscription]
