+++ pubsub/subscriptions/test_permissions.go
@@ -12,10 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Sample pubsub-quickstart creates a Google Cloud Pub/Sub topic.
-package main
+// Package subscription is a tool to manage Google Cloud Pub/Sub subscriptions by using the Pub/Sub API.
+// See more about Google Cloud Pub/Sub at https://cloud.google.com/pubsub/docs/overview.
+package subscription
 
-// [START pubsub_quickstart_create_topic]
 import (
 	"context"
 	"fmt"
@@ -24,28 +24,21 @@ import (
 	"cloud.google.com/go/pubsub"
 )
 
-func main() {
+func testPermissions(c *pubsub.Client, subName string) ([]string, error) {
 	ctx := context.Background()
 
-	// Sets your Google Cloud Platform project ID.
-	projectID := "YOUR_PROJECT_ID"
-
-	// Creates a client.
-	client, err := pubsub.NewClient(ctx, projectID)
+	// [START pubsub_test_subscription_permissions]
+	sub := c.Subscription(subName)
+	perms, err := sub.IAM().TestPermissions(ctx, []string{
+		"pubsub.subscriptions.consume",
+		"pubsub.subscriptions.update",
+	})
 	if err != nil {
-		log.Fatalf("Failed to create client: %v", err)
+		return nil, fmt.Errorf("TestPermissions: %v", err)
 	}
-
-	// Sets the name for the new topic.
-	topicName := "my-topic"
-
-	// Creates the new topic.
-	topic, err := client.CreateTopic(ctx, topicName)
-	if err != nil {
-		log.Fatalf("Failed to create topic: %v", err)
+	for _, perm := range perms {
+		log.Printf("Allowed: %v", perm)
 	}
-
-	fmt.Printf("Topic %v created.\n", topic)
+	// [END pubsub_test_subscription_permissions]
+	return perms, nil
 }
-
-// [END pubsub_quickstart_create_topic]
