+++ pubsub/topics/list_topics.go
@@ -12,40 +12,38 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Sample pubsub-quickstart creates a Google Cloud Pub/Sub topic.
-package main
+// Package topics is a tool to manage Google Cloud Pub/Sub topics by using the Pub/Sub API.
+// See more about Google Cloud Pub/Sub at https://cloud.google.com/pubsub/docs/overview.
+package topics
 
-// [START pubsub_quickstart_create_topic]
+// [START pubsub_list_topics]
+
 import (
 	"context"
 	"fmt"
-	"log"
 
 	"cloud.google.com/go/pubsub"
+	"google.golang.org/api/iterator"
 )
 
-func main() {
+func list(client *pubsub.Client) ([]*pubsub.Topic, error) {
 	ctx := context.Background()
 
-	// Sets your Google Cloud Platform project ID.
-	projectID := "YOUR_PROJECT_ID"
+	var topics []*pubsub.Topic
 
-	// Creates a client.
-	client, err := pubsub.NewClient(ctx, projectID)
-	if err != nil {
-		log.Fatalf("Failed to create client: %v", err)
+	it := client.Topics(ctx)
+	for {
+		topic, err := it.Next()
+		if err == iterator.Done {
+			break
+		}
+		if err != nil {
+			return nil, fmt.Errorf("Next: %v", err)
+		}
+		topics = append(topics, topic)
 	}
 
-	// Sets the name for the new topic.
-	topicName := "my-topic"
-
-	// Creates the new topic.
-	topic, err := client.CreateTopic(ctx, topicName)
-	if err != nil {
-		log.Fatalf("Failed to create topic: %v", err)
-	}
-
-	fmt.Printf("Topic %v created.\n", topic)
+	return topics, nil
 }
 
-// [END pubsub_quickstart_create_topic]
+// [END pubsub_list_topics]
