+++ pubsub/topics/publish_scale.go
@@ -1,0 +1,67 @@
+// Copyright 2019 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// Package topics is a tool to manage Google Cloud Pub/Sub topics by using the Pub/Sub API.
+// See more about Google Cloud Pub/Sub at https://cloud.google.com/pubsub/docs/overview.
+package topics
+
+// [START pubsub_publish_with_error_handling_that_scales]
+
+import (
+	"context"
+	"fmt"
+	"log"
+	"strconv"
+	"sync"
+	"sync/atomic"
+
+	"cloud.google.com/go/pubsub"
+)
+
+func publishThatScales(client *pubsub.Client, topic string, n int) error {
+	ctx := context.Background()
+	var wg sync.WaitGroup
+	var totalErrors uint64
+	t := client.Topic(topic)
+
+	for i := 0; i < n; i++ {
+		result := t.Publish(ctx, &pubsub.Message{
+			Data: []byte("Message " + strconv.Itoa(i)),
+		})
+
+		wg.Add(1)
+		go func(i int, res *pubsub.PublishResult) {
+			defer wg.Done()
+			// The Get method blocks until a server-generated ID or
+			// an error is returned for the published message.
+			id, err := res.Get(ctx)
+			if err != nil {
+				// Error handling code can be added here.
+				log.Output(1, fmt.Sprintf("Failed to publish: %v", err))
+				atomic.AddUint64(&totalErrors, 1)
+				return
+			}
+			fmt.Printf("Published message %d; msg ID: %v\n", i, id)
+		}(i, result)
+	}
+
+	wg.Wait()
+
+	if totalErrors > 0 {
+		return fmt.Errorf("%d of %d messages did not publish successfully", totalErrors, n)
+	}
+	return nil
+}
+
+// [END pubsub_publish_with_error_handling_that_scales]
