+++ pubsub/topics/publish_single.go
@@ -1,0 +1,43 @@
+// Copyright 2019 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// Package topics is a tool to manage Google Cloud Pub/Sub topics by using the Pub/Sub API.
+// See more about Google Cloud Pub/Sub at https://cloud.google.com/pubsub/docs/overview.
+package topics
+
+// [START pubsub_publisher_concurrency_control]
+import (
+	"context"
+	"fmt"
+
+	"cloud.google.com/go/pubsub"
+)
+
+func publishSingleGoroutine(client *pubsub.Client, topic string, msg []byte) error {
+	ctx := context.Background()
+	t := client.Topic(topic)
+	t.PublishSettings.NumGoroutines = 1
+
+	result := t.Publish(ctx, &pubsub.Message{Data: msg})
+	// Block until the result is returned and a server-generated
+	// ID is returned for the published message.
+	id, err := result.Get(ctx)
+	if err != nil {
+		return fmt.Errorf("Get: %v", err)
+	}
+	fmt.Printf("Published a message; msg ID: %v\n", id)
+	return nil
+}
+
+// [END pubsub_publisher_concurrency_control]
